+++ src/update.rs
@@ -238,6 +238,11 @@ fn update_move(
             let mode = Mode::Move(index);
             Model { state, mode }
         }
+        MoveMsg::Demote => {
+            let (state, index) = state.demote(index);
+            let mode = Mode::Move(index);
+            Model { state, mode }
+        }
         MoveMsg::Done => Model { state, mode: Mode::Select(index) },
     };
     Command::None(model)