+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+## [2.10.0-alpha] - 2025-01-20
+
+### Added
+- **Promote** command to move a subtree up a level.
+- **Demote** command to move a subtree down a level.
+
 ## [2.9.0-alpha] - 2025-01-13
 
 ### Removed
@@ -113,7 +119,8 @@
   - **Edit**: Modify item text.
   - **Delete**: Delete an item.
 
-[Unreleased]: https://github.com/darthkeith/sieve-selector/compare/v2.9.0-alpha...HEAD
+[Unreleased]: https://github.com/darthkeith/sieve-selector/compare/v2.10.0-alpha...HEAD
+[2.10.0-alpha]: https://github.com/darthkeith/sieve-selector/compare/v2.9.0-alpha...v2.10.0-alpha
 [2.9.0-alpha]: https://github.com/darthkeith/sieve-selector/compare/v2.8.0-alpha...v2.9.0-alpha
 [2.8.0-alpha]: https://github.com/darthkeith/sieve-selector/compare/v2.7.0-alpha...v2.8.0-alpha
 [2.7.0-alpha]: https://github.com/darthkeith/sieve-selector/compare/v2.6.0-alpha...v2.7.0-alpha